foo2

+++ vxsort/defs.h
@@ -51,7 +51,7 @@ namespace vxsort {
 
     using i8  = int8_t;
     using u8  = uint8_t;
-    using i16 = int32_t;
+    using i16 = int16_t;
     using i32 = int32_t;
     using i64 = int64_t;
     using u16 = uint16_t;
